render/stages: Move FromLinear source file
Rename the apply_transfer source to from_linear to match the name
change.
