+++ jxl/src/render/stages/mod.rs
@@ -3,9 +3,9 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-mod apply_transfer;
 mod chroma_upsample;
 mod convert;
+mod from_linear;
 mod gaborish;
 mod nearest_neighbor;
 mod noise;
